Modyfikacja Roli department boss i dodadanie podstawowej funkcji workera

+++ src/main/java/com/perfect/hepdeskapp/documentation/DocumentationController.java
@@ -3,11 +3,15 @@ package com.perfect.hepdeskapp.documentation;
 import com.perfect.hepdeskapp.attachment.Attachment;
 import com.perfect.hepdeskapp.attachment.AttachmentRepository;
 import com.perfect.hepdeskapp.config.FileUploadService;
+import com.perfect.hepdeskapp.department.Department;
 import com.perfect.hepdeskapp.ticket.Ticket;
 import com.perfect.hepdeskapp.ticket.TicketRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import javax.print.Doc;
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -24,12 +28,13 @@ public class DocumentationController {
     final
     AttachmentRepository attachmentRepository;
 
+
     public DocumentationController(DocumentationRepository documentationRepository, TicketRepository ticketRepository, AttachmentRepository attachmentRepository) {
         this.documentationRepository = documentationRepository;
         this.ticketRepository = ticketRepository;
         this.attachmentRepository = attachmentRepository;
     }
-    @PostMapping("/api/addDocumentation")
+    @PostMapping("/manager/api/addDocumentation")
     public String addDocumentation(@RequestParam("ticket") Long ticket, @RequestParam("content") String docContent,
      @RequestParam(value = "attachments", required = false) MultipartFile[] files) throws IOException {
         Documentation documentation = new Documentation();
@@ -57,30 +62,64 @@ public class DocumentationController {
         documentationRepository.saveAndFlush(documentation);
         return "redirect:/t/"+ticket;
     }
-    @PostMapping("/api/editDocumentation")
-    public String editDocumentation(@RequestParam("documentationid") Long documentationId, @RequestParam("ticket")  Long ticket, @RequestParam("content") String docContent,
+    @PostMapping("/manager/api/editDocumentation")
+    public String editDocumentation(@RequestParam("documentationid") Long documentationId, @RequestParam("ticketid")  Long ticket, @RequestParam("content") String docContent,
                                     @RequestParam(value = "attachments", required = false) MultipartFile[] files) throws IOException{
-        Documentation documentation = documentationRepository.findDocumentationById(documentationId);
-        documentation.setDescription(docContent);
-
-        documentationRepository.saveAndFlush(documentation);
+       Documentation documentation = documentationRepository.findDocumentationById(documentationId);
+       documentation.setDescription(docContent);
+       if(files!=null){
+           List<Attachment> attachments = new ArrayList<>();
+           Attachment attachment = attachmentRepository.findAttachmentByDocumentation(documentation.getId()).get(0);
+           if(attachment!=null){
+               String result = attachment.getUrl();
+               String[] r = result.split("/");
+               Path path = Paths.get("src","main","webapp","WEB-INF","uploads",r[2]);
+               String uploadDir = path + "/";
+               for(MultipartFile file : files){
+                   if(!file.isEmpty()){
+                       FileUploadService.saveFile(uploadDir,file.getOriginalFilename(),file);
+                       attachments.add(new Attachment(file.getOriginalFilename(),"/uploads/"+r[2]+"/"+file.getOriginalFilename()));
+                   }
+               }
+               attachmentRepository.saveAllAndFlush(attachments);
+               documentation.getDocumentationAttachmentsSet().addAll(attachments);
+           }
+       }
+       documentationRepository.saveAndFlush(documentation);
         return "redirect:/t/"+ticket;
     }
 
-    @GetMapping("/api/deleteDocumentation/{documentationId}")
+    @GetMapping("/manager/api/deleteDocumentation/{documentationId}")
     public String deleteDocumentation(@PathVariable("documentationId") Long documentation){
         Documentation documentationObject = documentationRepository.findDocumentationById(documentation);
         Long ticket = documentationObject.getTicket().getId();
-        if(documentationObject.getDocumentationAttachmentsSet() != null || !documentationObject.getDocumentationAttachmentsSet().isEmpty())
+        if(documentationObject.getDocumentationAttachmentsSet() != null && !documentationObject.getDocumentationAttachmentsSet().isEmpty())
         {
             Attachment attachment = attachmentRepository.findAttachmentByDocumentation(documentationObject.getId()).get(0);
             String result = attachment.getUrl();
             String[] r = result.split("/");
             Path path = Paths.get("src","main","webapp","WEB-INF","uploads",r[2]);
             FileUploadService.deleteDirectory(path.toFile());
+
         }
         documentationRepository.delete(documentationObject);
         return "redirect:/t/"+ticket;
+    }
+    @PostMapping("/manager/api/deleteAttachmentFromDocumentation")
+    @ResponseBody
+    public String deleteAttachmentFromDocumentation(@RequestParam("documentation_id") Long documentation_id, @RequestParam("attachment_id") Long attachment_id){
+        Documentation documentation = documentationRepository.findDocumentationById(documentation_id);
+        documentation.deleteAttachmentFromDocumentation(attachmentRepository.getReferenceById(attachment_id));
+        Attachment attachment = attachmentRepository.findAttachmentByDocumentation(documentation.getId()).get(0);
+        String result = attachment.getUrl();
+        String[] r = result.split("/");
+        Path path = Paths.get("src","main","webapp","WEB-INF","uploads",r[2]);
+        String uploadDir = path + "/"+attachmentRepository.getAttachmentById(attachment_id).getName();
+        File file = new File(uploadDir);
+        FileUploadService.deleteFile(file);
+        documentationRepository.saveAndFlush(documentation);
+        attachmentRepository.delete(attachmentRepository.getAttachmentById(attachment_id));
+        return "Successfully";
     }
